requirements.txt
pyTelegramBotAPI==4.2.0
python-decouple==3.5
requests==2.26.0
selenium==4.0.0